+++ src/app.py
@@ -35,10 +35,9 @@ main_task_page = st.Page("pages/main_task_page.py")
 stepwise_training_page = st.Page("pages/stepwise_training_page.py")
 task_1_content_page = st.Page("pages/task_1_content_page.py")
 task_2_content_page = st.Page("pages/task_2_content_page.py")
+feedback_task_2_page = st.Page("pages/feedback_task_2_page.py")
 task_3_content_page = st.Page("pages/task_3_content_page.py")
 feedback_task_3_page = st.Page("pages/feedback_task_3_page.py")
-task_4_content_page = st.Page("pages/task_4_content_page.py")
-feedback_task_4_page = st.Page("pages/feedback_task_4_page.py")
 
 pg = st.navigation([welcome_page,
                     consent_page,
@@ -57,7 +56,6 @@ pg = st.navigation([welcome_page,
                     task_2_content_page,
                     task_3_content_page,
                     feedback_task_3_page,
-                    task_4_content_page,
-                    feedback_task_4_page],position = "hidden")
+                    feedback_task_2_page],position = "hidden")
 pg.run()
