+++ src/pages/experiment_step_2_page.py
@@ -50,7 +50,7 @@ def feedback_page(text_container_1, feedback_container_1, progr_cont_1, text_con
     text_container_1.markdown(f"**Original statement:** {current_ori_statement}")
 
     feedback_container_1.markdown(
-        f"The AI classifies this statement as **{'Truthful' if ori_classification == 1 else 'Deceptive'}**.\n"
+        f"The AI classifies this statement as **{'TRUTHFUL' if ori_classification == 1 else 'DECEPTIVE'}**.\n"
         f"Confidence Score: **{ori_score:.2f}%**"
     )
     progr_cont_1.progress(int(ori_score)) 
@@ -58,11 +58,11 @@ def feedback_page(text_container_1, feedback_container_1, progr_cont_1, text_con
     text_container_2.markdown(f"**Your statement:** {current_repharsed_text}")
 
     feedback_container_2.markdown(
-        f"The AI classifies this statement as **{'Truthful' if paraphrase_classification == 1 else 'Deceptive'}**.\n"
+        f"The AI classifies this statement as **{'TRUTHFUL' if paraphrase_classification == 1 else 'DECEPTIVE'}**.\n"
         f"Confidence Score: **{classification_score:.2f}%**"
     )
     progr_cont_2.progress(int(classification_score))
-    opposite_classification = 'Deceptive' if ori_classification == 1 else 'Truthful'
+    opposite_classification = 'DECEPTIVE' if ori_classification == 1 else 'TRUTHFUL'
     # Conditionally display text_container_3 after "Retry" is clicked
     if 'submit_view' in st.session_state and st.session_state['submit_view'] == 1:
         text_container_3.markdown(
