+++ src/pages/instructions_page.py
@@ -4,7 +4,7 @@ import uuid
 st.title("Instructions")
 st.write(":book: In this experiment, you will read **3** short statements about past experiences. Each statement is either truthful :white_check_mark: or deceptive :lying_face:. The statements have been shortened and might at times end aprubtly.")
 st.write("Apart from the statements, we provide you with the predictions of a state-of-the-art lie detection algorithm based on artificial intelligence (AI) :robot_face:.")
-st.write("This shows you whether the AI thinks the statement is a deceptive or a truthful and how confident it is in its classification (Confidence Score).")
+st.write("This shows you whether the AI thinks the statement is deceptive or truthful and how confident it is in its classification (Confidence Score).")
 
 st.write("Your task is to **paraphrase** these statements. Specifically, your rewrite is meant to make the statement appear opposite in credibility compared to how they were originally classified by the AI (deceptive or truthful).")
 st.write("In total you will have **10** attempts per statement to decrease the **confidence score** of the AI's original prediction as much as possible. If you manage to change the original prediction (truthful or deceptive), you will immediately move on to the next statement.")
